browser/base/content/test/siteIdentity/test_no_mcb_on_http_site_font2.html

<!DOCTYPE HTML>
<html>
<!--
  Test 3 for Bug 909920 - See file browser_no_mcb_on_http_site.js for description.
  https://bugzilla.mozilla.org/show_bug.cgi?id=909920
-->
<head>
  <meta charset="utf-8">
  <title>Test 3 for Bug 909920</title>
  <link rel="stylesheet" type="text/css" href="https://example.com/browser/browser/base/content/test/siteIdentity/test_no_mcb_on_http_site_font2.css" />
<script src="/tests/SimpleTest/SimpleTest.js"></script>

<script type="text/javascript">
  async function checkLoadStates() {
   let state = await SpecialPowers.getSecurityState(window);

   var loadedMixedActive =
     !!(state & SpecialPowers.Ci.nsIWebProgressListener.STATE_LOADED_MIXED_ACTIVE_CONTENT);
   is(loadedMixedActive, false, "OK: Should not load mixed active content!");

   var blockedMixedActive =
     !!(state & SpecialPowers.Ci.nsIWebProgressListener.STATE_BLOCKED_MIXED_ACTIVE_CONTENT);
   is(blockedMixedActive, false, "OK: Should not block mixed active content!");

   var loadedMixedDisplay =
     !!(state & SpecialPowers.Ci.nsIWebProgressListener.STATE_LOADED_MIXED_DISPLAY_CONTENT);
   is(loadedMixedDisplay, false, "OK: Should not load mixed display content!");

   var blockedMixedDisplay =
     !!(state & SpecialPowers.Ci.nsIWebProgressListener.STATE_BLOCKED_MIXED_DISPLAY_CONTENT);
   is(blockedMixedDisplay, false, "OK: Should not block mixed display content!");

   var newValue = "Verifying MCB does not trigger warning/error for an http page ";
    newValue += "with https css that imports another http css which includes http font";
   document.getElementById("testDiv").innerHTML = newValue;
  }
</script>
</head>
<body onload="checkLoadStates()">
  <div class="testDiv" id="testDiv">
    Testing MCB does not trigger warning/error for an http page with https css that imports another http css which includes http font
  </div>
</body>
</html>
